requirements.txt
keras~=2.3.1
numpy~=1.19.1
gym~=0.17.2

matplotlib~=3.3.1
pillow~=7.2.0


keyboard~=0.13.5
